refactor: moved components folder
